pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Production-grade RAG Chatbot"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.3",
    "uvicorn[standard]>=0.42.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.11",
    "alembic>=1.18.4",
    "pydantic-settings>=2.0.0",
    "langchain>=0.3.0",
    "langchain-mistralai>=0.2.0",
    "langchain-qdrant>=0.2.0",
    "qdrant-client>=1.12.0",
    "python-multipart>=0.0.9",
    "sse-starlette>=2.0.0",
    "pypdf>=5.0.0",
    "fastembed>=0.4.0",
    "langchain-text-splitters>=1.1.1",
    "datasets>=4.8.4",
    "pillow>=12.2.0",
    "httpx>=0.28.1",
    "tenacity>=9.1.4",
]
